bond: Fix broken rebalancing after link state changes.
There are 3 constraints for moving hashes from one slave to another:

	1. The load difference is larger than ~3% of one slave's load.
	2. The load difference between slaves exceeds 100000 bytes.
	3. Moving of the hash makes the load difference lower by > 10%.

In current implementation if one of the slaves goes DOWN state, all
the hashes assigned to it will be moved to other slaves. After that,
if slave will go UP it will wait for rebalancing to get some hashes.
But in case where we have more than 10 equally loaded hashes it
will never fit constraint #3, because each hash will handle less than
10% of the load. Situation become worse when number of flows grows
higher and it's almost impossible to migrate any hash when all the
256 hash entries are used which is very likely when we have few
hundreds/thousands of flows.

As a result, if one of the slaves goes down and up while traffic
flows, it will never be used again for packet transmission.
Situation will not be fixed even if we'll stop traffic completely
and start it again because first two constraints will block
rebalancing on the earlier stages while we have low amount of traffic.

Moving of one hash if destination has no hashes as it was before
commit c460a6a7bc75 ("ofproto/bond: simplify rebalancing logic")
will not help because having one hash isn't enough to make load
difference less than 10% of total load and this slave will
handle only that one hash forever.

To fix this lets try to move few hashes simultaniously to fit
constraint #3.

Implementation includes sorting of 'entries' to be able to collect
entries with accumulated load close enough to ideal value.

Signed-off-by: Ilya Maximets <i.maximets@samsung.com>