dpdk: Unify passing of DPDK EAL arguments.
This patch deprecates the usage of current configuration knobs for
DPDK EAL arguments:

  - dpdk-alloc-mem
  - dpdk-socket-mem
  - dpdk-socket-limit
  - dpdk-lcore-mask
  - dpdk-hugepage-dir
  - dpdk-extra

All above configuration options replaced with single 'dpdk-options'
which has same format as old 'dpdk-extra', i.e. just a string with
all the DPDK EAL arguments.

All the documentation about old configuration knobs removed. Users
could still use an old interface, but the deprecation warning will be
printed. In case 'dpdk-options' provided, values of old options will
be ignored. New users will start using new 'dpdk-options' as this is
the only documented way providing EAL arguments.

Rationale:

From release to release DPDK becomes more complex. New EAL arguments
appears, old arguments becomes deprecated. Some changes their meaning
and behaviour. It's not easy task to manage all this and current
code in OVS that tries to manage DPDK options is not flexible/extendable
enough even to track all the dependencies between options in DPDK 18.11.
For example, '--socket-mem' changed its meaning, '--legacy-mem' and
'--socket-limit' appeared. But '--legacy-mem' and '--socket-limit'
could not be used at the same time and, also, we want to provide
good default value for '--socket-limit' to keep the consistent
behaviour of OVS memory usage. And this will be worse in the future.

Another point is that even now we have mutually exclusive database
configs in OVS and we have to handle them. i.e we're providing
'dpdk-alloc-mem' and 'dpdk-socket-mem' at the same time. Additionally,
user allowed to configure same options in 'dpdk-extra'. So, we have
similar/equal options in a three places in ovsdb and we need to
choose which one to use. It's pretty easy for user to get into
inconsistent database configuration, because users could update
one field without removing others, and OVS has to resolve all the
conflicts somehow constructing the EAL args. But we do not know in
practice, if the resulted arguments are the arguments that user wanted
to see or just forget to remove the higher priority knob.

Next point is that we have 'dpdk-lcore-mask' (-c) which is actually not
so user-friendly as '-l', but we have no option for it. Will we create
additional 'dpdk-lcore-list' ? I guess, no, because it's not really
important thing. But users will stuck with this not so user-friendly
masks.

Another thing is that OVS is not really need to check the consistency
of the EAL arguments because DPDK could check it instead. DPDK will
always check the args and it will do it better. There is no real
need to duplicate this functionality.

Keeping all the options in a single string 'dpdk-options' allows:

  * To have only one place for all the configs, i.e. it will be harder
    for user to forget to remove something from the single string while
    updating it.
  * Not to check the consistency between different configs, DPDK could
    check the cmdline itself.
  * Not to document every DPDK EAL argument in OVS. We can just
    describe few of them and point to DPDK docs for more information.
  * OVS still able to provide some minimal default arguments.
    Thanks to DPDK 18.11 we only need to specify an lcore. All other
    arguments are not necessary. (We still also providing default
    --socket-limit in case --socket-mem passed without it and without
    --legacy-mem)

Usage example:

  ovs-vsctl set Open_vSwitch . \
      other_config:dpdk-options="-l 1 -n 4 --socket-mem 1024,1024"

Signed-off-by: Ilya Maximets <i.maximets@samsung.com>