ovsdb-data: Optimize comparison for equality of db atoms.
ovsdb_datum_equals() used heavily in conditional monitoring and also
comparison of ovsdb atoms for equality is a frequent enough operation.
However, it's implemented via 3way 'strcmp'-like comparison that does
a lot of unnecessary operations trying to check 'lexicographical' order
of elements while it is not needed if we only need to know if values
are equal.

Implementation of a specific 'equals' function should speed up
ovsdb-server operations.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>