memory: Allow memory trimming if consumption grew significantly.
Many parts of different libopenvswitch-based applications uses lots of
small memory allocations and these are mostly cached by libc and not
released back to the system if freed by the application.  For example,
ovsdb-based applications like ovsdb-server or any application that
receives lots of data via ovsdb-idl (ovn-controller, ovn-northd)
allocates and frees lots of ovsdb_datum objects that contains many
small chunks of dynamically allocated memory.  While in case of
ovsdb-server it's easy to find a good place to manually trim the
heap (ovsdb compaction is a very good point in time to do that), for
ovsdb-idl based applications it's not obvious at all.

This change allows to enable memory trimming if memory monitoring
detects significant growth of consumption.  This should allow
applications to keep their memory usage in a meaningful range.

Added general-purpose 'memory/trim-on-growth on|off' unixctl command
to enable the feature.  It will be available to any application that
uses memory_run().  Disable by default to avoid possible performance
issues since the next allocation after trimming will end up in a
syscall.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>