ovsdb: Add support for transaction forwarding to the replication mode.
Current version of ovsdb replication allows to scale out read-only
access to the primary database.  However, many clients are not
read-only but read-mostly.  For example, ovn-controller.

In order to scale out database access for this case ovsdb-server
need to process transactions that are not read-only.  Replica is
not allowed to do that, i.e. not allowed to modify the database,
but it can act like a proxy and forward transactions that includes
database modifications to the primary server and forward replies
back to client. At the same time it may serve read-only transactions
and monitor requests by itself greatly reducing the load on primary
server.

To support that new command line option '--enable-txn-forward' added
to ovsdb-server.

This configuration will slightly increase transaction latency, but
it's not very important for read-mostly use cases.

With this change instead of creating a trigger to commit the
transaction, ovsdb-server will create a trigger for transaction
forwarding.  Later, replication_run() will send all new transactions
to the replication source.  Once transaction reply received from the
replication source, replication module will update the state of the
transaction forwarding with the reply.  After that, trigger_run()
will complete the trigger and jsonrpc_server_run() will send the
reply back to the client.  Since transaction reply from the replication
source will be received after all the updates, client will receive
all the updates before receiving the transaction reply as it is in
a normal scenario without transaction forwarding.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>