replication: Allow replication of _Server database.
If replication server is connected to a cluster member, clients
would like to know the state of the cluster from that server.
This will allow them to make a decision about re-connection.

Marking 'Database' table with a 'copyForReplication' flag to
have a new '_synced_Database' table where records from the
active server will be stored.  This way client will see state
of local databases in 'Database' table and state of databases
of the active server in '_synced_Database' table.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>