ovsdb: transaction: Incremental reassessment of weak refs.
The main idea is to not store list of weak references in the source
row, so they all don't need to be re-checked/updated on every
modification of that source row.  The point is that source row already
knows UUIDs of all destination rows stored in the data, so there is no
much profit in storing this information somewhere else.  If needed,
destination row can be looked up and reference can be looked up in the
destination row.  For the fast lookup, destination row now stores
references in a hash map.

Weak reference structure now contains the table and uuid of a source
row instead of a direct pointer.  This allows to replace/update the
source row without breaking any weak references stored in destination
rows.

Structure also now contains the key-value pair of atoms that triggered
creation of this reference.  These atoms can be used to quickly
subtract removed references from a source row.  During reassessment,
ovsdb now only needs to care about new added or removed atoms, and
atoms that got removed due to removal of the destination rows, but
these are marked for reassessment by the destination row.

Results of an OVN load-balancer benchmark:

  make sandbox SANDBOXFLAGS="--nbdb-model=clustered"
  pkill ovn-controller ovn-northd ovn-ic
  ./lb-stress.sh 120 3000

Before:

  %CPU  CPU Time    RSS    CMD
  58.6  00:07:10  2610012  ovsdb-server nb1.db
  34.9  00:04:16  2589604  ovsdb-server nb2.db
  34.7  00:04:14  2590928  ovsdb-server nb3.db
  35.6  00:04:12    74124  ovn-nbctl daemon

After:

  %CPU  CPU Time    RSS    CMD
  26.4  00:02:05  2703908  ovsdb-server nb1.db
  10.4  00:00:49  2683972  ovsdb-server nb2.db
  10.1  00:00:48  2684572  ovsdb-server nb3.db
  61.0  00:03:52    74312  ovn-nbctl daemon

So, on a cluster leader the processing time dropped by 3.44x, on
followers - by 5.2x.  I'm not exactly sure why nbctl took slightly
less time.  Probably, because ovsdb-server was faster and it spent
less time waiting on it, or something like that.  There is a slight
increase of memory usage, because new reference structure is larger.
Can, probably, be optimized.

According to 'perf', the main time consuming part of the new
implementation is construction of the 'added' and 'removed' sets for
the old/new datum pair.  This is because the implementation is not
100% incremental.  ovsdb_datum_added_removed() is still linear even
though it's fast.  And we actually have the same thing for strong
references and also for monitor updates, where ovsdb creates a diff
between old and new before sending transaction result to monitors or
writing on disk.  At the same time, this is not a problem of ovsdb
logic, but a technical issue that can be resolved.  E.g., by
constructing 'added' and 'removed' sets while applying the diff from
the transaction json, i.e. inside the ovsdb_datum_diff_apply_in_place,
and carrying it around and using instead of re-constructing each time.
This way we should be able to avoid any linear operations on sets.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>