uuid: Use memcmp to compare for equality.
memcmp() can not be used for 3-way comparison, because we want to
sort UUIDs in a way as they will be sorted in their string
representation, and memcmp() will not do that on LE systems.
However, it's perfectly fine to use it while checking for equality.

Current version of a code typically translated by compiler to
something like this:

    mov a->parts[0], reg32
    cmp b->parts[0], reg32
    je B
 A:
    <code that should be executed if not equal>
    jmp C
 B:
    mov a->parts[1], reg32
    cmp b->parts[1], reg32
    jne A
    <same 3 commands for parts 2 and 3>
 C:
    <code for 'equal' case>

This code is optimized for != case, i.e. the function inlined in a way
that code that should be executed if not equal is placed in a good
position and it's expected that first comparison will fail.

At the same time, the most performance critical parts for UUID
comparison are hash lookups, e.g. database row lookup by the reference,
or comparisons of old and new datum in the transaction.
In these scenarios equality is much more likely, and execution takes
the least performant path in the code snippet above.

Replacement with memcmp() leads to generation of a following code:

    mov a->parts[0-1], reg64_1
    mov a->parts[2-3], reg64_2
    xor b->parts[0-1], reg64_1
    xor b->parts[2-3], reg64_2
    or reg64_1, reg64_2
    je A
    <code that should be executed if equal>
 A:
    <code for 'not equal' case>

This much better suites the case of equality and gives a slight, but
measurable performance boost to database row lookups.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>