netdev-offload-tc: Use masks instead of keys while parsing tunnel attributes.
If the key is zero, it doesn't mean we don't need to match on it.
Masks should be checked instead.

For the metadata length that's a bit tricky, because the length in the
mask section of the 'flower' structure represents the actual length of
the mask and not the mask of the length field in the key.  But lengths
of the mask and the key must be equal in the OVS use case, so we can
use them interchangeably.

Fixes: 49a7961fca65 ("lib/tc: Avoid matching on tunnel ttl or tos if not needed")
Fixes: a468645c6d33 ("lib/tc: add geneve with option match offload")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>