hmap: Prefetch next entry on full iteration.
While using HMAP_FOR_EACH[_SAFE] macros, users will likely access
all the elements in the hash map.  So, it makes sense to prefetch
the next entry.

For a hash map search _WITH_HASH/_IN_BUCKET variants of these macros
are used and they use different access functions.  So, this use
case is not covered, and it shouldn't be.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>