ovsdb: Fix database statistics during the database replacement.
The counter for the number of atoms has to be re-set to the number from
the new database, otherwise the value will be incorrect.  For example,
this is causing the atom counter doubling after online conversion of
a clustered database.

Miscounting may also lead to increased memory consumption by the
transaction history or otherwise too aggressive transaction history
sweep.

Fixes: 317b1bfd7dd3 ("ovsdb: Don't let transaction history grow larger than the database.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>