ovsdb-tool: Fix cluster-to-standalone for DB conversion records.
If database conversion happens, both schema and the new data are
present in the database record.  However, the schema is just silently
ignored by ovsdb-tool cluster-to-standalone.  This creates data
inconsistency if the new data contains new columns, for example, so
the resulting database file will not be readable, or data will be lost.

Fix that by re-setting the database whenever a conversion record is
found and actually writing a new schema that will match the actual
data.  The database file will not be that similar to the original,
but there is no way to represent conversion in a standalone database
file format otherwise.

Fixes: 00de46f9ee42 ("ovsdb-tool: Convert clustered db to standalone db.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>